requirements.txt
africastalking==1.2.3
amqp==5.0.5
anyjson==0.3.3
asgiref==3.3.1
astroid==2.3.1
billiard==3.6.3.0
cairocffi==1.2.0
CairoSVG==2.5.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
contextlib2==0.6.0.post1
cssselect2==0.4.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.5
django-ckeditor==5.8.0
django-countries==7.1
django-crispy-forms==1.8.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-multiselectfield==0.1.12
django-notifications-hq==1.6.0
django-pwa==1.0.10
django-template-maths==0.2.0
django-timezone-field==4.1.1
django-user-activity-log2==0.0.24
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.0.0
html5lib==1.1
idna==2.10
isort==4.3.21
jdcal==1.4.1
jsonfield==3.1.0
kombu==5.0.2
lazy-object-proxy==1.4.2
Markdown==3.3.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.1.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.2
PyPDF2==1.26.0
Pyphen==0.10.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2019.3
PyYAML==5.4.1
reportlab==3.5.59
requests==2.25.1
schema==0.7.4
segno==1.3.1
six==1.12.0
sqlparse==0.3.0
swapper==1.1.2.post1
tablib==3.0.0
tinycss2==1.1.0
twilio==6.53.0
typed-ast==1.4.0
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==52.2
webencodings==0.5.1
whitenoise==4.1.4
wrapt==1.11.2
xlrd==2.0.1
xlwt==1.3.0
